bulk-upload-service/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
httpx
python-dotenv